refactor: apply cargo clippy --fix mechanical lints across 95 files
- derives/impl Default di mana new() tanpa default
- io::Error::other, redundant closures, needless_late_init, div_ceil,
  is_multiple_of, sort_by_key, iter().flatten(), matches! dll
- hapus impl Default JITEvaluator duplikat (conflict E0119 dari --fix)
- semua machine-applicable suggestion, tanpa perubahan semantik

+++ crates/mivon-elaboration/src/util/const_fold.rs
@@ -484,7 +484,7 @@ pub fn try_fold_const(
     if !const_expr_is_signed(expr) {
         if let Some(val) = try_fold_const_unsigned(expr, params)? {
             let width = const_fold_width(expr, params).unwrap_or_else(|| {
-                let abs = val as u64;
+                let abs = val;
                 let min_width = if val == 0 {
                     1
                 } else {
